Move plugin validation into plugins service
Delete the stale internal/plugininvocation package and expose plugin validation from services/plugins. Callers now use pluginservice.ValidateEffectiveCatalog(s) and pluginservice.ValidateDependencies while behavior stays unchanged.

Interface diff:

- plugininvocation.ValidateEffectiveCatalogsAndDependencies(ctx, cfg)

+ pluginservice.ValidateEffectiveCatalogsAndDependencies(ctx, cfg)

- plugininvocation.ValidateDependencies(ctx, cfg)

+ pluginservice.ValidateDependencies(ctx, cfg)

Example:

pluginservice.ValidateEffectiveCatalog(ctx, name, entry)
