+++ gestaltd/internal/daemon/provider.go
@@ -14,10 +14,10 @@ import (
 	"strings"
 
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/plugininvocation"
 	"github.com/valon-technologies/gestalt/server/internal/pluginsource"
 	"github.com/valon-technologies/gestalt/server/internal/providerpkg"
 	providermanifestv1 "github.com/valon-technologies/gestalt/server/sdk/providermanifest/v1"
+	pluginservice "github.com/valon-technologies/gestalt/server/services/plugins"
 	"gopkg.in/yaml.v3"
 )
 
@@ -414,7 +414,7 @@ func validateStagedReleaseCatalog(staged *providerpkg.StagedPreparedInstall) err
 	if err != nil {
 		return fmt.Errorf("invalid source in staged manifest: %w", err)
 	}
-	return plugininvocation.ValidateEffectiveCatalog(context.Background(), src.PluginName(), &config.ProviderEntry{
+	return pluginservice.ValidateEffectiveCatalog(context.Background(), src.PluginName(), &config.ProviderEntry{
 		ResolvedManifestPath: staged.ManifestPath,
 		ResolvedManifest:     staged.Manifest,
 	})
